Use Polytope class and related stuff only if Lemon is available
