+++ test/objectfeatures/test.cxx
@@ -1490,171 +1490,6 @@ struct AccumulatorTest
             shouldEqual(W(3, 0, 1), get<AutoRangeHistogram<3> >(c,3));
         }
     }
-
-    void testConvexHullFeatures2D()
-    {
-        using namespace vigra::acc;
-        std::string prefix("testConvexHullFeatures2D(): ");
-
-        int size = 6;
-        MultiArray<2, int> mask(vigra::Shape2(size, size));
-
-        mask(1, 1) = 1;
-        mask(2, 1) = 1;
-        mask(2, 2) = 1;
-        mask(2, 3) = 1;
-        mask(1, 3) = 1;
-        mask(3, 1) = 1;
-        mask(3, 3) = 1;
-        mask(4, 1) = 1;
-        mask(4, 3) = 1;
-
-        AccumulatorChainArray<
-                CoupledArrays<2, int>, 
-                Select<LabelArg<1>, ConvexHullFeatures> > chf;
-        chf.ignoreLabel(0);
-        extractFeatures(mask, chf);
-
-        getAccumulator<ConvexHullFeatures>(chf, 1).finalize();
-
-        {
-            TinyVector<double, 2> ref(2.5 - 1./18., 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).inputCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).inputCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        {
-            TinyVector<double, 2> ref(2.5, 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).hullCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).hullCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).inputVolume(), 9);
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).hullVolume(), 12);
-        {
-            TinyVector<double, 2> ref(8. / 3., 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).defectCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).defectCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).defectCount(), 2);
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeMean(),
-                1.5,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeVariance(),
-                0.5,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeSkewness(),
-                0.0,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeKurtosis(),
-                0.0,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectDisplacementMean(),
-                ((2.5 - 1./18. - 1.) + 2*(3.5 - 2.5 + 1./18.))/3.,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).convexity(),
-                9. / 12.,
-                (std::numeric_limits<double>::epsilon() * 2));
-    }
-
-    void testConvexHullFeatures3D()
-    {
-        using namespace vigra::acc;
-        std::string prefix("testConvexHullFeatures3D(): ");
-
-        int size = 5;
-        MultiArray<3, int> mask(vigra::Shape3(size, size, size), 0);
-        for (int i = 0; i < 9; i++)
-        {
-            mask(i / 3 + 1, i % 3 + 1, 1) = 1;
-            mask(i / 3 + 1, i % 3 + 1, 3) = 1;
-        }
-        mask(3, 1, 2) = 1; mask(3, 2, 2) = 1;
-        mask(1, 3, 2) = 1; mask(2, 3, 2) = 1;
-
-        // z = 0   | z = 1   | z = 2   | z = 3
-        // --------+---------+---------+--------
-        // 0 0 0 0 | 0 0 0 0 | 0 0 0 0 | 0 0 0 0
-        // 0 0 0 0 | 0 x x x | 0 0 0 x | 0 x x x
-        // 0 0 0 0 | 0 x x x | 0 0 0 x | 0 x x x
-        // 0 0 0 0 | 0 x x x | 0 x x 0 | 0 x x x
-
-        AccumulatorChainArray<
-                CoupledArrays<3, int>,
-                Select<LabelArg<1>, ConvexHullFeatures> > chf;
-        chf.ignoreLabel(0);
-        extractFeatures(mask, chf);
-
-        getAccumulator<ConvexHullFeatures>(chf, 1).finalize();
-        {
-            // x and y coordinate: (7*1 + 7*2 + 8*3) / 22 = 45 / 22
-            TinyVector<double, 3> ref(45. / 22., 45. / 22., 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).inputCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).inputCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        {
-            TinyVector<double, 3> ref(2., 2., 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).hullCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).hullCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).inputVolume(), 22);
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).hullVolume(), 27);
-        {
-            // x and y coordinate: (2*1 + 2*2 + 1*3) / 5 = 9 / 5
-            TinyVector<double, 3> ref(9. / 5., 9. / 5., 2.);
-            shouldEqualSequenceTolerance(
-                    get<ConvexHullFeatures>(chf, 1).defectCenter().begin(),
-                    get<ConvexHullFeatures>(chf, 1).defectCenter().end(),
-                    ref.begin(),
-                    (std::numeric_limits<double>::epsilon() * 2));
-        }
-        shouldEqual(get<ConvexHullFeatures>(chf, 1).defectCount(), 2);
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeMean(),
-                2.5,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeVariance(),
-                9. / 2.,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeSkewness(),
-                0.0,
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectVolumeKurtosis(),
-                0.0,
-                (std::numeric_limits<double>::epsilon() * 2));
-        // (sqrt(2) * 4 * (45 / 22 - 3 / 2) + sqrt(2) * 1 * (3 - 45 / 22)) / 5
-        // = sqrt(2) / 5 * (90 / 11 - 3 - 45 / 22)
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).defectDisplacementMean(),
-                sqrt(2.) / 5. * (90. / 11. - 3. - 45. / 22.),
-                (std::numeric_limits<double>::epsilon() * 2));
-        shouldEqualTolerance(
-                get<ConvexHullFeatures>(chf, 1).convexity(),
-                22. / 27.,
-                (std::numeric_limits<double>::epsilon() * 2));
-    }
 };
 
 struct FeaturesTestSuite : public vigra::test_suite
@@ -1671,8 +1506,6 @@ struct FeaturesTestSuite : public vigra::test_suite
         add(testCase(&AccumulatorTest::testHistogram));
         add(testCase(&AccumulatorTest::testRegionAccumulators));
         add(testCase(&AccumulatorTest::testIndexSpecifiers));
-        add(testCase(&AccumulatorTest::testConvexHullFeatures2D));
-        add(testCase(&AccumulatorTest::testConvexHullFeatures3D));
     }
 };
